// frontend/Types/dataTypes.ts
class Stats {
  XP: number = 0;
  GP: number = 0;
  rank: number = 0;
  numGames: number = 0;
  gamesWon: number = 0;
}

export class rankObj {
  rank: string = allRanks[4];
  tier: any = rank4;
}

export class UserTypeNew {
  id: number = 0;
  login: string = "";
  avatar: string = "";
  fullname: string = "";
  stats: Stats = new Stats();
  rank: number = 0;
  GP: number = 0;
  numGames: number = 0;
  gamesWon: number = 0;
  achievement: number[] = [];
  status: string = "";
  relation: string = "";
  constructor() {}
}

export const EmtyUser: UserTypeNew = {
  id: 0,
  login: "",
  avatar: "",
  fullname: "",
  stats: new Stats(),
  rank: 0,
  GP: 0,
  numGames: 0,
  gamesWon: 0,
  achievement: [],
  status: "",
  relation: "",
};